docs: fix stale references — PUBLISH.md rewrite, eval README dead rows, drop dead demo proof tool
- PUBLISH.md: the 0.4.0 launch one-shots (publish, propagation, unpublish) are
  executed — registry shows 0.7.0 and @vendoai/cli 404s. Rewritten around what
  is still live: steady-state releases, the OPEN @vendoai/harnesses first-publish
  blocker, trusted-publishing config, clean-room verify.
- docs/eval/README.md: 4 of 9 rows pointed into the scrubbed-and-gitignored
  docs/verification/ tree; folded into history pointers.
- demo-bank .vendo: host_proof_list synced tool pointed at the MAPLE_HARNESS
  proof route, which 404s in the shipped demo; dropped from tools.json and
  overrides.json.
