+++ .github/workflows/perf.yml
@@ -12,24 +12,6 @@ jobs:
   perf:
     runs-on: ubuntu-latest
 
-    services:
-      postgres:
-        image: postgres:16
-        env:
-          POSTGRES_USER: vendo
-          POSTGRES_PASSWORD: vendo
-          POSTGRES_DB: vendo_bench
-        ports:
-          - 5432:5432
-        options: >-
-          --health-cmd "pg_isready -U vendo"
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-
-    env:
-      POSTGRES_URL: postgres://vendo:vendo@localhost:5432/vendo_bench
-
     steps:
       - uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4
       - uses: pnpm/action-setup@b906affcce14559ad1aafd4ab0e942779e9f58b1 # v4
