+++ tools/genui-bench/fixtures/maple.ts
@@ -1,20 +1,105 @@
 /**
- * The Maple (demo-bank) host fixture: catalog/tools/shape cards come from the
- * shared bench loaders in @vendoai/apps (the same surface the live pipeline
- * harnesses generate against), the theme from examples/demo-bank/.vendo/theme.json,
- * and every cataloged tool gets a deterministic canned-data executor so the
- * rendered app's queries and actions run at view time. All money is integer
- * CENTS (the Maple law). No Math.random, no Date.now() — same call, same data.
+ * The Maple (demo-bank) host fixture: catalog + tools loaded from
+ * examples/demo-bank/.vendo/*.json (overrides.json merged, auth/demo/voice
+ * plumbing filtered out — the cadence loader pattern), hand-written shape
+ * cards mirroring examples/demo-bank/src/server/types.ts, the theme from
+ * examples/demo-bank/.vendo/theme.json, and every cataloged tool gets a
+ * deterministic canned-data executor so the rendered app's queries and
+ * actions run at view time. All money is integer CENTS (the Maple law).
+ * No Math.random, no Date.now() — same call, same data.
  */
 import { readFileSync } from "node:fs";
 import { fileURLToPath } from "node:url";
 import { dirname, resolve } from "node:path";
-import { VendoError, vendoThemeSchema } from "@vendoai/core";
-import { demoBankToolShapes, loadDemoBankCatalog, loadDemoBankTools } from "@vendoai/apps";
+import { VendoError, vendoThemeSchema, type NormalizedCatalog, type ShapeType } from "@vendoai/core";
+import type { HostToolInfo } from "@vendoai/apps";
 import type { HostFixture } from "../runner/types";
 
 const here = dirname(fileURLToPath(import.meta.url));
 const repoRoot = resolve(here, "../../..");
+const vendoDir = resolve(repoRoot, "examples/demo-bank/.vendo");
+
+const loadMapleCatalog = (): NormalizedCatalog => {
+  const raw = JSON.parse(readFileSync(resolve(vendoDir, "catalog.json"), "utf8")) as {
+    entries: Array<{ name: string; description: string; propsSchema: unknown; examples?: string[] }>;
+  };
+  return raw.entries.map((entry) => ({
+    name: entry.name,
+    description: entry.description,
+    propsJsonSchema: entry.propsSchema as NormalizedCatalog[number]["propsJsonSchema"],
+    ...(entry.examples === undefined ? {} : { examples: entry.examples }),
+  })) as NormalizedCatalog;
+};
+
+const loadMapleTools = (): HostToolInfo[] => {
+  const raw = JSON.parse(readFileSync(resolve(vendoDir, "tools.json"), "utf8")) as {
+    tools: Array<{ name: string; description: string; risk: string; inputSchema?: Record<string, unknown> }>;
+  };
+  // tools.json is what `vendo sync` writes; the live registry merges
+  // overrides.json on top (disabled, risk, description). Mirror that here or
+  // the fixture sees tools the deployment disabled.
+  const { tools: overrides = {} } = JSON.parse(
+    readFileSync(resolve(vendoDir, "overrides.json"), "utf8"),
+  ) as { tools?: Record<string, { disabled?: boolean; risk?: string; description?: string }> };
+  return raw.tools
+    .filter(({ name }) => overrides[name]?.disabled !== true)
+    .filter(({ name }) => !name.startsWith("host_auth") && !name.startsWith("host_demo") && !name.startsWith("host_voice"))
+    .map(({ name, description, risk, inputSchema }) => ({
+      name,
+      description: overrides[name]?.description ?? description,
+      risk: overrides[name]?.risk ?? risk,
+      ...(inputSchema === undefined ? {} : { inputSchema }),
+    }));
+};
+
+// ---------------------------------------------------------------------------
+// Shape cards — mirror examples/demo-bank/src/server/types.ts: what runtime
+// shape-sampling would derive from live responses.
+// ---------------------------------------------------------------------------
+
+const str: ShapeType = { kind: "string" };
+const num: ShapeType = { kind: "number" };
+const bool: ShapeType = { kind: "boolean" };
+const arr = (items: ShapeType): ShapeType => ({ kind: "array", items });
+const obj = (fields: Record<string, ShapeType>): ShapeType => ({ kind: "object", fields });
+
+const account = obj({
+  id: str, name: str, kind: str, mask: str, balance: num,
+  accountNumber: str, sparkline: arr(num),
+});
+const transaction = obj({
+  id: str, accountId: str, merchant: str, descriptor: str, amount: num,
+  timestamp: str, category: str, status: str, method: str,
+});
+const paged = (items: ShapeType): ShapeType => obj({ data: arr(items), total: num });
+
+export const mapleToolShapes: Record<string, ShapeType> = {
+  host_getProfile: obj({ name: str, email: str, netWorth: num, accountCount: num, avatarInitials: str }),
+  host_listAccounts: arr(account),
+  host_getAccount: account,
+  host_listTransactions: paged(transaction),
+  host_listAccountTransactions: paged(transaction),
+  host_getTransaction: transaction,
+  host_listCards: arr(obj({
+    id: str, accountId: str, type: str, network: str, mask: str,
+    expMonth: num, expYear: num, frozen: bool, design: str,
+  })),
+  host_listCardTransactions: paged(transaction),
+  host_getBudgets: arr(obj({ category: str, limit: num, spent: num })),
+  host_listGoals: arr(obj({ id: str, name: str, target: num, saved: num, icon: str })),
+  host_listPayees: arr(obj({ id: str, name: str, kind: str })),
+  host_listScheduledPayments: arr(obj({
+    id: str, payeeId: str, payeeName: str, amount: num, nextDate: str, cadence: str,
+  })),
+  host_getSpendingInsights: arr(obj({ category: str, amount: num })),
+  host_getCashflowInsights: arr(obj({ label: str, in: num, out: num })),
+  host_getRecurringInsights: arr(obj({
+    id: str, merchant: str, amount: num, cadence: str, category: str, nextDate: str,
+  })),
+  host_listNotifications: arr(obj({
+    id: str, kind: str, title: str, body: str, at: str, read: bool,
+  })),
+};
 
 // ---------------------------------------------------------------------------
 // Seed data — fixed constants mirroring examples/demo-bank/src/server/types.ts.
@@ -187,12 +272,10 @@ const executors: Record<string, Executor> = {
 
 export const mapleFixture: HostFixture = {
   name: "maple",
-  catalog: loadDemoBankCatalog(),
-  tools: loadDemoBankTools(),
-  shapes: demoBankToolShapes,
-  theme: vendoThemeSchema.parse(
-    JSON.parse(readFileSync(resolve(repoRoot, "examples/demo-bank/.vendo/theme.json"), "utf8")),
-  ),
+  catalog: loadMapleCatalog(),
+  tools: loadMapleTools(),
+  shapes: mapleToolShapes,
+  theme: vendoThemeSchema.parse(JSON.parse(readFileSync(resolve(vendoDir, "theme.json"), "utf8"))),
   async execute(tool, input) {
     const executor = executors[tool];
     if (executor === undefined) throw new VendoError("not-found", `maple fixture has no tool "${tool}"`);
